requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
